# drop "sstream" dependency