@Totktonada 2021-03-05
> compiled on centos:7 but run on centos:8.

I hope that this should be considered as improper use of the build/pack machinery, at least in the general case.