@rosik 2021-03-05
Before:
<details> <summary>master, 4dba1943</summary>
# Test transcoding performance
    # fixtures
    1..3
    # FIXTURE 'ebay'
    # size: 34.73 KiB
    # decode: 13849.46 Req/s
    # encode: 12512.76 Req/s
    ok - reencode 'ebay'
    # FIXTURE 'reed'
    # size: 277.01 KiB
    # decode: 345.25 Req/s
    # encode: 403.22 Req/s
    ok - reencode 'reed'
    # FIXTURE 'customer'
    # size: 503.57 KiB
    # decode: 216.93 Req/s
    # encode: 270.25 Req/s
    ok - reencode 'customer'
    # Bandwidth (decode average): 223.24 MiB/s
    # Bandwidth (encode average): 222.10 MiB/s
    # fixtures: end
ok - fixtures
</details>


After:
<details> <summary>v2, 60f92ee3</summary>
# Test transcoding performance
    # fixtures
    1..3
    # FIXTURE 'ebay'
    # size: 34.73 KiB
    # decode: 13986.48 Req/s
    # encode: 15588.19 Req/s
    ok - reencode 'ebay'
    # FIXTURE 'reed'
    # size: 277.01 KiB
    # decode: 325.57 Req/s
    # encode: 700.08 Req/s
    ok - reencode 'reed'
    # FIXTURE 'customer'
    # size: 503.57 KiB
    # decode: 207.25 Req/s
    # encode: 490.28 Req/s
    ok - reencode 'customer'
    # Bandwidth (decode average): 221.37 MiB/s
    # Bandwidth (encode average): 319.70 MiB/s
    # fixtures: end
ok - fixtures
</details>


| Test            | master |    v2 |     change |
|-----------------|--------|-------|------------|
| ebay decode     |  13849 | 13986 |  1.0098924 |
| ebay encode     |  12512 | 15588 |  1.2458440 |
| read decode     |    345 |   325 | 0.94202899 |
| read encode     |    403 |   700 |  1.7369727 |
| customer decode |    216 |   207 | 0.95833333 |
| customer encode |    270 |   490 |  1.8148148 |
|-----------------|--------|-------|------------|
| decode average  |    223 |   221 | 0.99103139 |
| encode average  |    222 |   319 |  1.4369369 |