@olegrok 2021-03-05
> I hope that this should be considered as improper use of the build/pack machinery, at least in the general case.

It's true. But seems nobody allows me to remove static link with system libraries - it could cause some problems when we deliver binary rocks to our customers.